pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "pyarrow (>=19.0.0,<20.0.0)",
    "xlsxwriter (>=3.2.0,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "google-genai (>=0.5.0,<0.6.0)",
    "google-generativeai (>=0.8.3,<0.9.0)",
    "nodriver (>=0.39,<0.40)",
    "lxml (>=5.3.0,<6.0.0)",
    "html5lib (>=1.1,<2.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pymupdf4llm (>=0.0.17,<0.0.18)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faiss-cpu (>=1.9.0.post1,<2.0.0)",
    "pydantic-ai @ git+https://github.com/santokalayil/pydantic-ai.git@sandev",
    "crawl4ai (>=0.4.248,<0.5.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "google-cloud-aiplatform (>=1.79.0,<2.0.0)",
    "playwright (>=1.50.0,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

